R13 navbar updating (#138)
* Updated the NavBar

Changed the logo and the links. Currently, the lins just go to the home page until those components are built out.

* Updated the NavBar

Changed the logo and the links. Currently, the lins just go to the home page until those components are built out.

* Changed vertical alignment

changed the alignment so that the buttons and logo were matched.

* Changed vertical alignment

changed the alignment so that the buttons and logo were matched.

* Changed vertical alignment

changed the alignment so that the buttons and logo were matched.

* updated snapshot

* Changed logo color

* Changed logo color