R13 navbar testing (#141)
* Added Navigation Tests

Added a couple tests for NavBar. Writing more soon, pausing to fix merge conflict on another branch

* Adding another test

* Updated Navigation Testing

WIP Added multiple tests for the Nav Bar when the user is not logged in including testing the logo link, the lifecylce functions. Once the routes and components have been added, and we have determined how the navbar will look for logged in users and admins, we can finish writing tests.

* WIP Updated Navigation Testing

WIP Added multiple tests for the Nav Bar when the user is not logged in including testing the logo link, the lifecylce functions. Once the routes and components have been added, and we have determined how the navbar will look for logged in users and admins, we can finish writing tests.

* pulled master so tests would pass