R13 footer (#143)
* Created Footer WIP

Created footer with newletter email form, social media links and logo. Expect logo to be updated. Social media links currently go to ask an expert as placeholder. Need to fix up css.

* Created Footer Frontend and Backend WIP

Created footer component, routes, and DB. Currently working on tests

* Created Footer Frontend and Backend

Finished component, API, routes, and DB. Testing is a little odd, but we decided to move on for now as we will be refactoring later.

* Ran Tests and Added Snapshots

Resolved merge conflict of missing snapshot in Home folder

* Reran Tests to Update Snapshot

Snapshot was failing so ran tests again to update.

* Fixed Footer Test

Fixed test to only find footer so that we snapshots will not run into router history issue

* Created Footer Backend Tests and Existing Email Catch

Wrote backend tests for footer newsletter sign up post and wrote a catch for if an email already exists in the database

* Created Footer Backend Tests and Existing Email Catch

Wrote backend tests for footer newsletter sign up post and wrote a catch for if an email already exists in the database

* Improved Newsletter Routes and Added Tests

Added a find single email function and added tests

* Completed all Change Requests Except One

Test for frontend Newsletter Form is still wonky, but everything else has been addressed

* Completed all Change Requests Except One

Test for frontend Newsletter Form is still wonky, but everything else has been addressed

* Added Last Test

* Added Last Test

* Fixed Backend  Newsletter Test, Frontend Newsletter Form and elevateAPI

* Updated Footer CSS

* Updated Footer CSS

* add padding above footer