Nav to sass (#164)
* refactored navigation component

Refactored to have three consts for LoggedIn, LoggedOut and Admin. Makes it easier to style individually. Also changed to react bootstrap, and was able to fixthe issue where narrowing the screen would mess up the navbar. Moved the admin button to the drop down.

* refactored navigation component

Refactored to have three consts for LoggedIn, LoggedOut and Admin. Makes it easier to style individually. Also changed to react bootstrap, and was able to fixthe issue where narrowing the screen would mess up the navbar. Moved the admin button to the drop down.

* Made some CSS changes

Made css changes to the drop down to make it look better and get rid of a magic bullet point. Made it so that when you click on a dropdown item, it does not turn blue.

* Made some CSS changes

Made css changes to the drop down to make it look better and get rid of a magic bullet point. Made it so that when you click on a dropdown item, it does not turn blue.

* Eliminated hard refresh and

Eleminated the hard refresh that was added when we made the switch to react bootstrap. Also made the dropdown items bold when hoovering over them. Other minor css

* Eliminated hard refresh and

Eleminated the hard refresh that was added when we made the switch to react bootstrap. Also made the dropdown items bold when hoovering over them. Other minor css

* Removed useless code and..

Removed useless code and made some transition changes and css changes to size of drop down.

* Removed useless code and..

Removed useless code and made some transition changes and css changes to size of drop down.

* trying to center everything

* trying to center everything

* Got the mobile view to work

* Got the mobile view to work

* small route change to avoid conflict

* Updated a test and deleted two

Deleted the component did mount and unmount tests because we got rid of the refs and that functionality

* Updated a test and deleted two

Deleted the component did mount and unmount tests because we got rid of the refs and that functionality

* Changed Dropdown

Gave the buttons borders. Made it so there were no warnings in the console about buried components. That fix made active buttons disappear which was fixed as well. Deleted alot of old ref references and code that was not being used.

* Changed Dropdown

Gave the buttons borders. Made it so there were no warnings in the console about buried components. That fix made active buttons disappear which was fixed as well. Deleted alot of old ref references and code that was not being used.

* Changed Dropdown

Gave the buttons borders. Made it so there were no warnings in the console about buried components. That fix made active buttons disappear which was fixed as well. Deleted alot of old ref references and code that was not being used.

* fixed package.json

* Fixed mobile view

* Updated Snapshot

* converted to sass

* converted to sass

* deleted css file